% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomass.R
\name{biomass}
\alias{biomass}
\title{Calculating biomass by species by life stage for each year}
\usage{
biomass(data, Species, Stage, Year, Mass)
}
\arguments{
\item{data}{data.frame to use}

\item{Species}{first column to group by}

\item{Stage}{second column to group by}

\item{Year}{third column to group by}

\item{Mass}{column that is summed}
}
\value{
data frame with biomass for each life stage of a species for each year
}
\description{
Calculating biomass by species by life stage for each year
}
