% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsedates.R
\name{parse_dates}
\alias{parse_dates}
\title{Creating a year and month column from dates for easy grouping and visualization}
\usage{
parse_dates(data)
}
\arguments{
\item{data}{data.frame to use}
}
\value{
same data.frame but now with a column that just has Year for each entry
}
\description{
Creating a year and month column from dates for easy grouping and visualization
}
