% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vis.R
\name{plot_data}
\alias{plot_data}
\title{Visualizing trends for long-term datasets}
\usage{
plot_data(data, Year, Biomass, Species)
}
\arguments{
\item{data}{data.frame to use}
}
\value{
plot with biomass by year with species colored
}
\description{
Visualizing trends for long-term datasets
}
