% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_HW6.R
\name{sum_g1g2}
\alias{sum_g1g2}
\title{Calculating sum of data grouped by two categories}
\usage{
sum_g1g2(data, group1, group2, tosum)
}
\arguments{
\item{data}{data.frame to use}

\item{group1}{first column to group by}

\item{group2}{second column to group by}

\item{tosum}{what to take the sum of}
}
\value{
data frame with sums of a numerical category grouped by two other categories
}
\description{
Calculating sum of data grouped by two categories
}
