% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_HW7.R
\name{yearify}
\alias{yearify}
\title{Creating a year column from dates for easy grouping and visualization}
\usage{
yearify(data)
}
\arguments{
\item{data}{data.frame to use}
}
\value{
same data.frame but now with a column that just has Year for each entry
}
\description{
Creating a year column from dates for easy grouping and visualization
}
